Move from a class interface to a function interface. 0.1.1 -> 0.2.0.
The original class/method design was intended as a way to configure the
dmenu prefences. However, partial application gets the job done as
documented in show's docstring. There's no state to keep track of if we're
willing to fetch the version message on every call. Type checking was removed to
make the code more "pythonic". The tests were removed since they really only
tested type checking.